requirements.txt
streamlit>=1.31.0
pypdf>=4.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0